ParmenidesChangeGame-js
Javascript code that allows user to two input any positive integer and game will return an object with key of change and value of a random message related to change.  Notes:  the second number is randomized and negative and non numbers are not allowed and user will be prompted to input a positve integer.  This code is easily executed by copy and pasting onto a console that supports Javascript.